+++ presto-main/src/test/java/com/facebook/presto/block/AbstractTestBlockCursor.java
@@ -24,7 +24,7 @@ import static org.testng.Assert.fail;
 
 public abstract class AbstractTestBlockCursor
 {
-    private final SortedMap<Long,Tuple> expectedValues = BlockCursorAssertions.toTuplesMap(createTestCursor());
+    private final SortedMap<Integer,Tuple> expectedValues = BlockCursorAssertions.toTuplesMap(createTestCursor());
 
     protected abstract Block createExpectedValues();
 
@@ -33,12 +33,12 @@ public abstract class AbstractTestBlockCursor
         return createExpectedValues().cursor();
     }
 
-    public final Tuple getExpectedValue(long position)
+    public final Tuple getExpectedValue(int position)
     {
         return getExpectedValues().get(position);
     }
 
-    public final SortedMap<Long, Tuple> getExpectedValues()
+    public final SortedMap<Integer, Tuple> getExpectedValues()
     {
         return expectedValues;
     }
@@ -47,7 +47,7 @@ public abstract class AbstractTestBlockCursor
     public void setUp()
     {
         // verify expected values
-        assertEquals(ImmutableList.copyOf(getExpectedValues().keySet()), ImmutableList.of(0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L));
+        assertEquals(ImmutableList.copyOf(getExpectedValues().keySet()), ImmutableList.of(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10));
     }
 
     @Test
@@ -95,7 +95,7 @@ public abstract class AbstractTestBlockCursor
         }
 
         // skip past end
-        assertFalse(cursor.advanceToPosition((long) 100));
+        assertFalse(cursor.advanceToPosition(100));
         assertTrue(cursor.isFinished());
         assertFalse(cursor.isValid());
     }
@@ -166,7 +166,7 @@ public abstract class AbstractTestBlockCursor
     {
         BlockCursor cursor = createTestCursor();
 
-        for (Entry<Long, Tuple> entry : getExpectedValues().entrySet()) {
+        for (Entry<Integer, Tuple> entry : getExpectedValues().entrySet()) {
             assertNextValue(cursor, entry.getKey(), entry.getValue());
         }
 
@@ -178,7 +178,7 @@ public abstract class AbstractTestBlockCursor
     {
         BlockCursor cursor = createTestCursor();
 
-        for (Entry<Long, Tuple> entry : getExpectedValues().entrySet()) {
+        for (Entry<Integer, Tuple> entry : getExpectedValues().entrySet()) {
             assertNextPosition(cursor, entry.getKey(), entry.getValue());
         }
 
@@ -190,7 +190,7 @@ public abstract class AbstractTestBlockCursor
     {
         BlockCursor cursor = createTestCursor();
 
-        for (Long position : getExpectedValues().keySet()) {
+        for (Integer position : getExpectedValues().keySet()) {
             assertNextValuePosition(cursor, position);
         }
 
@@ -203,8 +203,8 @@ public abstract class AbstractTestBlockCursor
     {
         BlockCursor cursor = createTestCursor();
 
-        for (Entry<Long, Tuple> entry : getExpectedValues().entrySet()) {
-            long position = entry.getKey();
+        for (Entry<Integer, Tuple> entry : getExpectedValues().entrySet()) {
+            int position = entry.getKey();
             Tuple tuple = entry.getValue();
             if (position % 2 != 0) {
                 assertNextValue(cursor, position, tuple);
